requirements.txt

Django==5.2.2
asgiref==3.8.1
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
cryptography==42.0.2 
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
#database info
# Database name: lockbox_db
# User: myrock
# Password: susant@5656
# Host: localhost
# Port: 5432

#  Fernet symmetric encryption 